# course3/2.server/6.API_go-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample weather app",
        "title": "WeatherApi",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/address/search": {
            "post": {
                "description": "Search coords by address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "SearchCoords",
                "parameters": [
                    {
                        "description": "House number, road, suburb, city, state, country",
                        "name": "coordinates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "house_number": {
                    "type": "string"
                },
                "road": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "suburb": {
                    "type": "string"
                }
            }
        },
        "main.RequestAddressSearch": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        }
    }
}